pyproject.toml
[tool.poetry]
name = "moot-api"
version = "1.0.0"
description = "The backend API for Moot"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.1"
python-dotenv = "^0.17.1"
uvicorn = "^0.13.4"
asyncpg = "^0.23.0"
itsdangerous = "^2.0.1"
aiohttp = "^3.7.4"
starlette-discord = "^0.1.1"
aioredis = "^1.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
